Refactor instance assignment logic to a separate class
